+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.0.12] - 2024-09-16
+### Added
+- Table component
+- Precompile nunjucks script (experimental)
+
+### Changed
+- Use `merge` filter in nunjucks
+
 ## [v0.0.11] - 2024-09-05
 ### Added
 - Tag component